[release][docs] Update documents for release of v0.7.0 (#124)

+++ benches/decimal128/bench_root.mojo
@@ -154,7 +154,7 @@ fn run_benchmark(
 
     # Calculate speedup factor (if Python benchmark ran)
     if python_time > 0:
-        var speedup = python_time / mojo_time
+        var speedup = Float64(python_time) / Float64(mojo_time)
         speedup_factors.append(Float64(speedup))
 
         # Print results with speedup comparison
